LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/GUM-impossible-constraint.https-expected.txt

When prompted, accept to share your video stream.

Description

This test checks that setting an impossible mandatory constraint (width >=1G) in getUserMedia works


FAIL Tests that setting an impossible constraint in getUserMedia fails assert_equals: An impossible constraint triggers a OverconstrainedError expected "OverconstrainedError" but got "Error"
